# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: bloodGroup
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- DOCTORS ---
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, name, age, gender, phone, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        age:         { type: integer }
        gender:      { type: string }
        bloodGroup:  { type: ["string", "null"] }
        address:     { type: ["string", "null"] }
        phone:       { type: string }
        email:       { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        diseaseHistory:   { type: ["string", "null"] }
        createdAt:   { type: string }
    PatientInput:
      type: object
      required: [name, age, gender, phone]
      properties:
        name:             { type: string, minLength: 1 }
        age:              { type: integer, minimum: 0 }
        gender:           { type: string }
        bloodGroup:       { type: string }
        address:          { type: string }
        phone:            { type: string }
        email:            { type: string }
        emergencyContact: { type: string }
        diseaseHistory:   { type: string }
    PatientUpdate:
      type: object
      properties:
        name:             { type: string }
        age:              { type: integer }
        gender:           { type: string }
        bloodGroup:       { type: string }
        address:          { type: string }
        phone:            { type: string }
        email:            { type: string }
        emergencyContact: { type: string }
        diseaseHistory:   { type: string }

    # --- Doctor ---
...